Skills/web_search/requirements.txt
flask 
requests
beautifulsoup4 
newspaper3k 
ddgs 
sentence-transformers 
transformers 
torch